# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlobeVote Civic Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/questions:
    get:
      operationId: listQuizQuestions
      tags: [quiz]
      summary: Get quiz questions by category and difficulty
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [voting-basics, electoral-systems, civic-rights, world-elections, misinformation]
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, expert]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
  /quiz/submit:
    post:
      operationId: submitQuizAnswer
      tags: [quiz]
      summary: Submit a quiz answer and get feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAnswerRequest"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAnswerResult"
  /quiz/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [quiz]
      summary: Get top players leaderboard
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, alltime]
            default: weekly
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # CLAIM VERIFICATION
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List curated civic news articles
      parameters:
        - name: countryId
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: News articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get a specific news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"

  # USER PROGRESS
  /progress/{sessionId}:
    get:
      operationId: getUserProgress
      tags: [progress]
      summary: Get user progress and civic skill level
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
  /progress/{sessionId}/achievements:
    get:
      operationId: getUserAchievements
      tags: [progress]
      summary: Get user achievements and badges
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"

  # STATS
...
components:
  schemas:
...
    QuizQuestion:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, expert]
        xpReward:
          type: integer
        explanation:
          type: string
      required: [id, question, options, category, difficulty, xpReward, explanation]
    QuizAnswerRequest:
      type: object
      properties:
        questionId:
          type: string
        selectedOption:
          type: integer
        sessionId:
          type: string
        timeTakenSeconds:
          type: integer
      required: [questionId, selectedOption, sessionId]
    QuizAnswerResult:
      type: object
      properties:
        correct:
          type: boolean
        correctOption:
          type: integer
        explanation:
          type: string
        xpEarned:
          type: integer
        streakBonus:
          type: integer
        newLevel:
          type: string
          nullable: true
      required: [correct, correctOption, explanation, xpEarned, streakBonus]
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        sessionId:
          type: string
        displayName:
          type: string
        level:
          type: string
        xp:
          type: integer
        questionsAnswered:
          type: integer
        accuracy:
          type: number
      required: [rank, sessionId, displayName, level, xp, questionsAnswered, accuracy]
...
    NewsArticle:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        source:
          type: string
        sourceUrl:
          type: string
        biasRating:
          type: string
          enum: [left, center-left, center, center-right, right, unknown]
        relevanceScore:
          type: number
        topic:
          type: string
        countryId:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        publishedAt:
          type: string
      required: [id, title, summary, source, sourceUrl, biasRating, relevanceScore, topic, publishedAt]
    UserProgress:
      type: object
      properties:
        sessionId:
          type: string
        displayName:
          type: string
        level:
          type: string
          enum: [beginner, civic-learner, analyst, policy-expert, democracy-champion]
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        questionsAnswered:
          type: integer
        correctAnswers:
          type: integer
        accuracy:
          type: number
        streak:
          type: integer
        countriesExplored:
          type: integer
        claimsVerified:
          type: integer
        chatInteractions:
          type: integer
        joinedAt:
          type: string
      required: [sessionId, displayName, level, xp, xpToNextLevel, questionsAnswered, correctAnswers, accuracy, streak, countriesExplored, claimsVerified, chatInteractions, joinedAt]
    Achievement:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          enum: [quiz, exploration, verification, chat, streak, special]
        unlockedAt:
          type: string
      required: [id, name, description, icon, category, unlockedAt]
...